Fix update account

See merge request 0_onnet_longnh/m_onnet_odoo_mfl2201er!3366

+++ addons/custom/forlife_invoice/models/invoice.py
@@ -221,7 +221,8 @@ class AccountMove(models.Model):
         tax_lines = self.line_ids.filtered(lambda x: x.display_type == 'tax').write({
             'tax_ids': False
         })
-        tax_lines.unlink()
+        if tax_lines:
+            tax_lines.unlink()
         self.sudo().write({
             'invoice_line_ids': False,
             'type_inv': type_po_cost if type_po_cost else False,
@@ -493,7 +494,8 @@ class AccountMove(models.Model):
                 tax_line_ids = rec.line_ids.filtered(lambda x: x.display_type == 'tax').write({
                     'tax_ids': False
                 })
-                tax_line_ids.unlink()
+                if tax_line_ids:
+                    tax_line_ids.unlink()
                 invoice_description = []
                 for vendor_back_id in rec.vendor_back_ids:
                     if vendor_back_id.invoice_description not in invoice_description:
